+++ helper/src/commonMain/kotlin/com/algolia/instantsearch/helper/relateditems/internal/extensions/MatchingPattern.kt
@@ -1,0 +1,32 @@
+package com.algolia.instantsearch.helper.relateditems.internal.extensions
+
+import com.algolia.instantsearch.helper.filter.state.FilterState
+import com.algolia.instantsearch.helper.filter.state.groupAnd
+import com.algolia.instantsearch.helper.filter.state.groupOr
+import com.algolia.instantsearch.helper.relateditems.MatchingPattern
+import com.algolia.instantsearch.helper.relateditems.internal.FilterFacetAndID
+import com.algolia.search.model.filter.Filter
+import com.algolia.search.model.filter.FilterGroupsConverter
+
+internal fun <T> List<MatchingPattern<T>>.toOptionalFilters(hit: T): List<List<String>>? {
+    val filterState = FilterState()
+    forEach { filterState.addMatchingPattern(hit, it) }
+    return FilterGroupsConverter.Legacy.Facet(filterState.toFilterFacetGroup()).unquote()
+}
+
+/**
+ * Create an [FilterFacetAndID] from a [MatchingPattern].
+ */
+internal fun <T> MatchingPattern<T>.toOptionalFilter(hit: T): FilterFacetAndID {
+    return when (val property = property.get(hit)) {
+        is Iterable<*> -> {
+            val groupOr = groupOr()
+            val list = property.map { value -> Filter.Facet(attribute, value.toString(), score) }.toTypedArray()
+            FilterFacetAndID(groupOr, list)
+        }
+        else -> FilterFacetAndID(
+            groupAnd(),
+            arrayOf(Filter.Facet(attribute, property.toString(), score))
+        )
+    }
+}